requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# LangGraph + LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0

# OpenAI (Whisper STT)
openai>=1.51.0

# ElevenLabs TTS (optional)
elevenlabs>=1.9.0

# HTTP client
httpx>=0.27.0

# Telegram Bot
python-telegram-bot>=21.0

# Email (Resend API)
resend>=2.0.0

# Utils
python-dotenv>=1.0.1
pydantic>=2.9.0


